Combine implementation of mods text drawable
I don't want to have to update things in multiple places with different
code in each place.

This also closes https://github.com/ppy/osu/issues/36412.
