+++ osu.Game/Screens/OnlinePlay/FooterButtonFreeModsV2.cs
@@ -5,7 +5,6 @@ using System;
 using System.Collections.Generic;
 using osu.Framework.Allocation;
 using osu.Framework.Bindables;
-using osu.Framework.Extensions.LocalisationExtensions;
 using osu.Framework.Graphics;
 using osu.Framework.Graphics.Containers;
 using osu.Framework.Graphics.Effects;
@@ -13,21 +12,19 @@ using osu.Framework.Graphics.Shapes;
 using osu.Framework.Graphics.Sprites;
 using osu.Framework.Utils;
 using osu.Game.Graphics;
-using osu.Game.Graphics.Sprites;
 using osu.Game.Localisation;
 using osu.Game.Overlays;
 using osu.Game.Overlays.Mods;
 using osu.Game.Rulesets.Mods;
 using osu.Game.Screens.Footer;
 using osu.Game.Screens.Play.HUD;
+using osu.Game.Screens.SelectV2;
 using osuTK;
 
 namespace osu.Game.Screens.OnlinePlay
 {
     public partial class FooterButtonFreeModsV2 : ScreenFooterButton
     {
-        private const float bar_height = 30f;
-
         public readonly Bindable<IReadOnlyList<Mod>> FreeMods = new Bindable<IReadOnlyList<Mod>>([]);
         public readonly Bindable<bool> Freestyle = new Bindable<bool>();
 
@@ -45,7 +42,7 @@ namespace osu.Game.Screens.OnlinePlay
         private Container modsWedge = null!;
         private ModDisplay modDisplay = null!;
         private Container modContainer = null!;
-        private ModCountText overflowModCountDisplay = null!;
+        private FooterButtonMods.ModCountText overflowModCountDisplay = null!;
 
         public FooterButtonFreeModsV2(ModSelectOverlay overlay)
             : base(overlay)
@@ -66,7 +63,7 @@ namespace osu.Game.Screens.OnlinePlay
                 Origin = Anchor.BottomLeft,
                 Shear = OsuGame.SHEAR,
                 CornerRadius = CORNER_RADIUS,
-                Size = new Vector2(BUTTON_WIDTH, bar_height),
+                Size = new Vector2(BUTTON_WIDTH, FooterButtonMods.BAR_HEIGHT),
                 Masking = true,
                 EdgeEffect = new EdgeEffectParameters
                 {
@@ -100,7 +97,7 @@ namespace osu.Game.Screens.OnlinePlay
                                 Current = { BindTarget = FreeMods },
                                 ExpansionMode = ExpansionMode.AlwaysContracted,
                             },
-                            overflowModCountDisplay = new ModCountText
+                            overflowModCountDisplay = new FooterButtonMods.ModCountText
                             {
                                 Mods = { BindTarget = FreeMods },
                                 Freestyle = { BindTarget = Freestyle }
@@ -144,57 +141,6 @@ namespace osu.Game.Screens.OnlinePlay
                 overflowModCountDisplay.Show();
             else
                 overflowModCountDisplay.Hide();
-        }
-
-        private partial class ModCountText : VisibilityContainer
-        {
-            public readonly Bindable<IReadOnlyList<Mod>> Mods = new Bindable<IReadOnlyList<Mod>>();
-            public readonly Bindable<bool> Freestyle = new Bindable<bool>();
-
-            private OsuSpriteText text = null!;
-
-            [Resolved]
-            private OverlayColourProvider colourProvider { get; set; } = null!;
-
-            protected override void LoadComplete()
-            {
-                base.LoadComplete();
-
-                RelativeSizeAxes = Axes.Both;
-
-                InternalChildren = new Drawable[]
-                {
-                    new Box
-                    {
-                        Colour = colourProvider.Background3,
-                        Alpha = 0.8f,
-                        RelativeSizeAxes = Axes.Both,
-                    },
-                    text = new OsuSpriteText
-                    {
-                        Anchor = Anchor.Centre,
-                        Origin = Anchor.Centre,
-                        Font = OsuFont.Torus.With(size: 14f, weight: FontWeight.Bold),
-                        Shear = -OsuGame.SHEAR,
-                    }
-                };
-
-                Mods.BindValueChanged(_ => updateText());
-                Freestyle.BindValueChanged(_ => updateText());
-
-                updateText();
-            }
-
-            protected override void PopIn() => this.FadeIn(300, Easing.OutExpo);
-            protected override void PopOut() => this.FadeOut(300, Easing.OutExpo);
-
-            private void updateText()
-            {
-                if (Freestyle.Value)
-                    text.Text = ModSelectOverlayStrings.AllMods.ToUpper();
-                else
-                    text.Text = ModSelectOverlayStrings.Mods(Mods.Value.Count).ToUpper();
-            }
         }
     }
 }
